// frontend/src/app/models/users.ts
export class User{
    id:             string;
    name:           string;
    surname:        string;
    username:       string;
    password:       string;
    email:          string;
    phone:          string;
    organization:   string;
    country:        string;
    city:           string;
    postalcode:     string;
    street:         string;
    type:           string;
    status:         string;
    photo:             any;
    taxNumber:      string;
  }
  